backend/requirements.txt
# EvoLabeler-Backend Requirements
# Generated from pyproject.toml

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic[email]==2.9.0
pydantic-settings==2.6.0

# Database & Storage
supabase==2.9.0

# Web Scraping
playwright==1.48.0

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
python-multipart==0.0.12
aiofiles==24.1.0
pillow==11.0.0

# Data Processing
pyyaml==6.0.2

# Development Dependencies
pytest==8.3.0
pytest-asyncio==0.24.0
black==24.10.0
ruff==0.7.0
mypy==1.13.0

